% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_matrices.R
\name{align_matrix_cols}
\alias{align_matrix_cols}
\title{Align the columns of two matrices}
\usage{
align_matrix_cols(x, y)
}
\arguments{
\item{x}{A matrix}

\item{y}{Another matrix}
}
\value{
A list with the input \code{x} and \code{y} matrices, with the
columns aligned using their names. Columns not in both matrices
are omitted.
}
\description{
Align the columns of two matrices using their column names,
omitting columns that are not present in both.
}
\examples{
# using the provided lineup2ex data (a list of two matrices)
# reduces to the common columns and puts the columns in the same order
# (using the column names)
aligned <- align_matrix_cols(lineup2ex$gastroc, lineup2ex$islet)

}
\seealso{
\code{\link[=align_matrix_rows]{align_matrix_rows()}}
}
