% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_matrices.R
\name{align_matrix_rows}
\alias{align_matrix_rows}
\title{Align the rows of two matrices}
\usage{
align_matrix_rows(x, y)
}
\arguments{
\item{x}{A matrix}

\item{y}{Another matrix}
}
\value{
A list with the input \code{x} and \code{y} matrices, with the rows
aligned using their names. Rows not in both matrices are
omitted.
}
\description{
Align the rows of two matrices using their row names, omitting rows
that are not present in both.
}
\examples{
y <- x <- matrix(nrow=3, ncol=2)
rownames(x) <- c("A", "B", "C")
rownames(y) <- c("B", "A", "D")
align_matrix_rows(x,y)

}
\seealso{
\code{\link[=align_matrix_cols]{align_matrix_cols()}}
}
