% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_betw_matrices.R
\name{dist_betw_matrices}
\alias{dist_betw_matrices}
\title{Distance between rows of two matrices}
\usage{
dist_betw_matrices(x, y, distance = c("rmsd", "mad"), cores = 1)
}
\arguments{
\item{x}{A numeric matrix.}

\item{y}{A second numeric matrix, with the same number of rows as \code{x}.}

\item{distance}{Indicates whether to use Euclidean distance
(\code{"rmsd"} for root mean square difference) or the mean absolute
difference (\code{"mad"}).}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel:detectCores]{parallel::detectCores()}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel:makeCluster]{parallel::makeCluster()}}.}
}
\value{
If \code{x} is \code{p} by \code{n} and \code{y} is \code{p} by \code{m}, then the result
is a \code{n} by \code{m} matrix whose (i,j)th element is the distance
between the ith column of \code{x} and the jth column of \code{y}.
}
\description{
Calculate the distances between the rows of one matrix and the rows of a second matrix.
}
\seealso{
\code{\link[=corr_betw_matrices]{corr_betw_matrices()}}
}
