% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonself.R
\name{get_nonself}
\alias{get_nonself}
\title{Get self-nonself distances}
\usage{
get_nonself(d)
}
\arguments{
\item{d}{A distance matrix}
}
\value{
The input distance matrix with all self-self distances
replaced with NAs.
}
\description{
Return the distance matrix with all self-self distances replaced
with NAs (and so just containing the self-self distances).
}
\seealso{
\code{\link[=get_self]{get_self()}}, \code{\link[=get_best]{get_best()}}, \code{\link[=get_2ndbest]{get_2ndbest()}}
}
