% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_problems.R
\name{get_problems}
\alias{get_problems}
\title{Summarize potential problems in a distance matrix}
\usage{
get_problems(d, dimension = c("row", "column"), get_min = TRUE,
  subset = c("problems", "all"))
}
\arguments{
\item{d}{A distance matrix}

\item{dimension}{Whether to determine the best distances within rows or columns}

\item{get_min}{If TRUE, get the minimum; if FALSE, get the maximum}

\item{subset}{Whether to return just the rows with potential problems, or all of the rows.}
}
\value{
A data frame containing individual ID, distance to self,
best distance and corresponding individual, 2nd best distance
and the corresponding individual.
}
\description{
For the inviduals represented in a distance matrix, collect the
self-self, best, and 2nd best distances, and summarize the results
in a data frame.
}
\seealso{
\code{\link[=get_self]{get_self()}}, \code{\link[=get_best]{get_best()}}, \code{\link[=get_2ndbest]{get_2ndbest()}}, \code{\link[=which_best]{which_best()}}, \code{\link[=get_nonself]{get_nonself()}}
}
