% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_self.R
\name{get_self}
\alias{get_self}
\title{Get self-self distance}
\usage{
get_self(d)
}
\arguments{
\item{d}{A distance matrix}
}
\value{
A vector with all of distinct individuals, with the self-self distance values.
}
\description{
For each individual represented in a distance matrix, find self-self entry (with NAs for individuals present in only the rows or only the columns.
}
