% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup2-package.R
\docType{package}
\name{lineup2-package}
\alias{lineup2}
\alias{lineup2-package}
\title{lineup2: Lining Up Two Sets of Measurements}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for detecting and correcting sample mix-ups between two sets of measurements, such as between gene expression data on two tissues. This is a revised version of the 'lineup' package, to be more general and not tied to the 'qtl' package.
}
\section{Vignette}{

\itemize{
\item \href{https://kbroman.org/lineup2/lineup2.html}{R/lineup2 User Guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/lineup2}
  \item Report bugs at \url{https://github.com/kbroman/lineup2/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

}
\keyword{internal}
