% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_best.R
\name{which_best}
\alias{which_best}
\title{Determine which individual has smallest distance to each individual}
\usage{
which_best(d, dimension = c("row", "column"), get_min = TRUE)
}
\arguments{
\item{d}{A distance matrix}

\item{dimension}{Whether to get the minimum by row or by column}

\item{get_min}{If TRUE, get the minimum; if FALSE, get the maximum}
}
\value{
A vector with \strong{all} distinct individuals, with the
character string labels for the individuals giving the minimum
(or maximum) value by row or column. We include all individuals
so that the results are aligned with the results of
\code{\link[=get_self]{get_self()}}.
}
\description{
For each individual represented in a distance matrix, find the
individual giving the smallest entry (with NAs for individuals
present in only the rows or only the columns).
}
\seealso{
\code{\link[=get_best]{get_best()}}, \code{\link[=get_self]{get_self()}}, \code{\link[=get_2ndbest]{get_2ndbest()}}, \code{\link[=which_2ndbest]{which_2ndbest()}}
}
