% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_matrices.R
\name{align_matrix_rows}
\alias{align_matrix_rows}
\title{Align the rows of two matrices}
\usage{
align_matrix_rows(x, y)
}
\arguments{
\item{x}{A matrix}

\item{y}{Another matrix}
}
\value{
A list with the input \code{x} and \code{y} matrices, with the rows
aligned using their names. Rows not in both matrices are
omitted.
}
\description{
Align the rows of two matrices using their row names, omitting rows
that are not present in both.
}
\examples{
# using the provided lineup2ex data (a list of two matrices)
# reduces to the common rows and puts the rows in the same order
# (using the row names)
aligned <- align_matrix_rows(lineup2ex$gastroc, lineup2ex$islet)

}
\seealso{
\code{\link[=align_matrix_cols]{align_matrix_cols()}}
}
