% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_betw_matrices.R
\name{dist_betw_matrices}
\alias{dist_betw_matrices}
\title{Distance between rows of two matrices}
\usage{
dist_betw_matrices(x, y)
}
\arguments{
\item{x}{A numeric matrix.}

\item{y}{A second numeric matrix, with the same number of columns as \code{x}.}
}
\value{
If \code{x} is \code{n} by \code{p} and \code{y} is \code{m} by \code{p}, then the result
is a \code{n} by \code{m} matrix whose (i,j)th element is the distance
between the ith row of \code{x} and the jth row of \code{y}.
}
\description{
Calculate the distances between the rows of one matrix and the rows of a second matrix.
}
