% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_self.R
\name{get_self}
\alias{get_self}
\title{Get self-self distance}
\usage{
get_self(d)
}
\arguments{
\item{d}{A distance matrix}
}
\value{
A vector with all distinct individuals, with the self-self values
}
\description{
For each individual represented in a distance matrix, pull the self-self entry (with NAs for individuals present in only the rows or only the columns).
}
\seealso{
\code{\link[=get_best]{get_best()}}, \code{\link[=get_2ndbest]{get_2ndbest()}}
}
