% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup2-package.R
\docType{package}
\name{lineup2-package}
\alias{lineup2}
\alias{lineup2-package}
\title{lineup2: Lining Up Two Sets of Measurements}
\description{
Tools for detecting and correcting sample mix-ups between two sets
of measurements, such as between gene expression data on two
tissues. This is a revised version of the lineup package, to be
more general.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/lineup2}
}

}
\keyword{internal}
